Changed max_frame_size argument of aws_cryptosdk_deserialize_frame from a uint64_t to a uint32_t

+++ include/aws/cryptosdk/private/framefmt.h
@@ -139,6 +139,6 @@ int aws_cryptosdk_deserialize_frame(
     /* in */
     struct aws_byte_cursor *ciphertext_buf,
     const struct aws_cryptosdk_alg_properties *alg_props,
-    uint64_t max_frame_size);
+    uint32_t max_frame_size);
 
 #endif
